pyproject.toml
[tool.poetry]
name = "openfabric-test"
version = "1.1"
description = "cognitive-assistant"
authors = ["Andrei"]

[tool.poetry.dependencies]
python = "^3.8"
openfabric-pysdk = "0.2.7"
transformers = "4.39.3"
tf-keras = "2.15.0"
torchvision= "0.17.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
